added addproductform component
